remove renaming changes done in api
