+++ tests/test_stainnorm.py
@@ -1,7 +1,7 @@
 """Tests for stain normalisation code."""
 
 from tiatoolbox.utils.misc import imread
-from tiatoolbox.tools.stain_norm import get_normaliser
+from tiatoolbox.tools.stainnorm import get_normaliser
 from tiatoolbox.tools import stainextract
 from tiatoolbox import cli
 from tiatoolbox.utils.exceptions import MethodNotSupported
@@ -165,7 +165,7 @@ def test_command_line_stainnorm(_source_image):
     stainnorm_result = runner.invoke(
         cli.main,
         [
-            "stain_norm",
+            "stain-norm",
             "--source_input",
             source_img,
             "--target_input",
@@ -180,7 +180,7 @@ def test_command_line_stainnorm(_source_image):
     stainnorm_result = runner.invoke(
         cli.main,
         [
-            "stain_norm",
+            "stain-norm",
             "--source_input",
             source_img,
             "--target_input",
@@ -195,7 +195,7 @@ def test_command_line_stainnorm(_source_image):
     stainnorm_result = runner.invoke(
         cli.main,
         [
-            "stain_norm",
+            "stain-norm",
             "--source_input",
             source_img,
             "--target_input",
@@ -210,7 +210,7 @@ def test_command_line_stainnorm(_source_image):
     stainnorm_result = runner.invoke(
         cli.main,
         [
-            "stain_norm",
+            "stain-norm",
             "--source_input",
             source_img,
             "--target_input",
@@ -232,7 +232,7 @@ def test_cli_stainnorm_dir(_source_image):
     stainnorm_result = runner.invoke(
         cli.main,
         [
-            "stain_norm",
+            "stain-norm",
             "--source_input",
             str(source_img),
             "--target_input",
@@ -254,7 +254,7 @@ def test_cli_stainnorm_file_not_found_error(_source_image):
     stainnorm_result = runner.invoke(
         cli.main,
         [
-            "stain_norm",
+            "stain-norm",
             "--source_input",
             str(source_img)[:-1],
             "--target_input",
@@ -278,7 +278,7 @@ def test_cli_stainnorm_method_not_supported(_source_image):
     stainnorm_result = runner.invoke(
         cli.main,
         [
-            "stain_norm",
+            "stain-norm",
             "--source_input",
             str(source_img),
             "--target_input",
