+++ tiatoolbox/tools/__init__.py
@@ -19,4 +19,4 @@
 # ***** END GPL LICENSE BLOCK *****
 
 """This package contains various tools for working with WSIs."""
-from tiatoolbox.tools import stain_norm, stainextract, patchextraction
+from tiatoolbox.tools import stainnorm, stainextract, patchextraction
